requirements.txt
Flask>=2.1.2
Flask_WTF>=1.0.1
numpy>=1.21.4
seldonian-engine
Werkzeug>=2.1.2
WTForms>=3.0.1
